mainProject/requirements.txt
3==0.0.0
asgiref==3.8.1
boto==2.49.0
boto3==1.34.157
botocore==1.34.157
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-environ==0.11.2
django-storages==1.14.4
djangorestframework==3.15.2
idna==3.7
jmespath==1.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
